Refactor and add support for new transitions to npec_comparison
