+++ src/evaluating_rewards/analysis/dissimilarity_heatmaps/plot_gridworld_heatmap.py
@@ -24,10 +24,11 @@ import numpy as np
 import pandas as pd
 import sacred
 
-from evaluating_rewards import serialize, tabular
+from evaluating_rewards import serialize
 from evaluating_rewards.analysis import gridworld_rewards, stylesheets, visualize
 from evaluating_rewards.analysis.dissimilarity_heatmaps import heatmaps, reward_masks
 from evaluating_rewards.analysis.reward_figures import gridworld_reward_heatmap
+from evaluating_rewards.distances import tabular
 from evaluating_rewards.scripts import script_utils
 
 plot_gridworld_heatmap_ex = sacred.Experiment("plot_gridworld_heatmap")
