+++ tests/test_epic_sample.py
@@ -24,12 +24,14 @@ import pytest
 from stable_baselines.common import vec_env
 import tensorflow as tf
 
-from evaluating_rewards import datasets, epic_sample, rewards, serialize, tabular
+from evaluating_rewards import datasets, serialize
 from evaluating_rewards import envs  # noqa: F401  # pylint:disable=unused-import
+from evaluating_rewards.distances import epic_sample, tabular
+from evaluating_rewards.rewards import base
 
 
 def mesh_evaluate_models_slow(
-    models: Mapping[epic_sample.K, rewards.RewardModel],
+    models: Mapping[epic_sample.K, base.RewardModel],
     obs: np.ndarray,
     actions: np.ndarray,
     next_obs: np.ndarray,
@@ -53,7 +55,7 @@ def mesh_evaluate_models_slow(
         dones=dones,
         infos=None,
     )
-    rews = rewards.evaluate_models(models, transitions)
+    rews = base.evaluate_models(models, transitions)
     rews = {k: v.reshape(len(obs), len(actions), len(next_obs)) for k, v in rews.items()}
     return rews
 
@@ -83,7 +85,7 @@ def test_mesh_evaluate_models(
         models = {}
         for i in range(n_models):
             with tf.variable_scope(str(i)):
-                models[i] = rewards.MLPRewardModel(space, space)
+                models[i] = base.MLPRewardModel(space, space)
 
         session.run(tf.global_variables_initializer())
         with session.as_default():
@@ -117,9 +119,9 @@ def test_sample_canon_shaping(
     with graph.as_default():
         with session.as_default():
             models = {k: serialize.load_reward(k, "dummy", venv, discount) for k in reward_types}
-            constant = rewards.ConstantReward(venv.observation_space, venv.action_space)
+            constant = base.ConstantReward(venv.observation_space, venv.action_space)
             constant.constant.set_constant(42.0)
-            models["big_sparse"] = rewards.LinearCombinationModelWrapper(
+            models["big_sparse"] = base.LinearCombinationModelWrapper(
                 {
                     "model": (
                         models["evaluating_rewards/PointMassSparseWithCtrl-v0"],
